refactor: moved http module to application/common module refactor: moved list-routes module to documentation module refactor: moved database module to application/common module
