+++ src/apps/backend/modules/access-token/types.ts
@@ -1,6 +1,6 @@
 import { PhoneNumber } from '../account/types';
 import { ApplicationError } from '../application';
-import { HttpStatusCodes } from '../http';
+import { HttpStatusCodes } from '../application/common/http';
 
 export class AccessToken {
   token: string;
