+++ src/apps/backend/modules/account/rest-api/account-controller.ts
@@ -1,5 +1,5 @@
 import { applicationController, Request, Response } from '../../application';
-import { HttpStatusCodes } from '../../http';
+import { HttpStatusCodes } from '../../application/common/http';
 import AccountService from '../account-service';
 import {
   Account,
