+++ src/apps/backend/modules/application/application-server.ts
@@ -1,10 +1,10 @@
 import bodyParser from 'body-parser';
 import express, { Application, NextFunction, Request, Response } from 'express';
 
-import { HttpStatusCodes } from '../http';
 import { Logger } from '../logger';
 
 import { ApplicationError } from './application-error';
+import { HttpStatusCodes } from './common/http';
 
 export enum ApplicationServerErrorCodes {
   UNHANDLED_ERROR = 'SERVER_ERR_01',
