+++ src/apps/backend/modules/application/common/database/objectid-utils.ts
