+++ src/apps/backend/modules/documentation/rest-api/documentation-controller.ts
@@ -1,5 +1,5 @@
 import { applicationController, Request, Response } from '../../application';
-import { HttpStatusCodes } from '../../http';
+import { HttpStatusCodes } from '../../application/common/http';
 import DocumentationService from '../documentation-service';
 
 import { serializeDocumentationAsJSON } from './documentation-serializer';
