+++ src/apps/backend/modules/password-reset-token/types.ts
@@ -1,5 +1,5 @@
 import { ApplicationError } from '../application';
-import { HttpStatusCodes } from '../http';
+import { HttpStatusCodes } from '../application/common/http';
 
 export class PasswordResetToken {
   id: string;
