+++ src/apps/backend/modules/task/rest-api/task-controller.ts
@@ -1,5 +1,5 @@
 import { applicationController, Request, Response } from '../../application';
-import { HttpStatusCodes } from '../../http';
+import { HttpStatusCodes } from '../../application/common/http';
 import TaskService from '../task-service';
 import {
   Task,
