+++ src/apps/backend/modules/task/types.ts
@@ -1,5 +1,5 @@
 import { ApplicationError } from '../application';
-import { HttpStatusCodes } from '../http';
+import { HttpStatusCodes } from '../application/common/http';
 
 export class Task {
   id: string;
