+++ test/spec/access-token/account-auth-middleware.spec.ts
@@ -12,8 +12,8 @@ import {
   accessAuthMiddleware,
 } from '../../../src/apps/backend/modules/access-token';
 import { Account } from '../../../src/apps/backend/modules/account';
+import { ObjectIdUtils } from '../../../src/apps/backend/modules/application/common/database';
 import ConfigService from '../../../src/apps/backend/modules/config/config-service';
-import { ObjectIdUtils } from '../../../src/apps/backend/modules/database';
 import { createAccount } from '../../helpers/account';
 
 describe('accessAuthMiddleware', () => {
